% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{arrange.SummarizedExperiment}
\alias{arrange.SummarizedExperiment}
\title{Arrange by variables}
\usage{
\method{arrange}{SummarizedExperiment}(.data, axis, ...)
}
\arguments{
\item{.data}{SummarizedExperiment to arrange}

\item{axis}{The axis to perform the operation on. Either row or col.}

\item{...}{Comma separated list of unquoted variable names, 
or expressions involving variable names. 
Use desc() to sort a variable in descending order}
}
\description{
Order either rows or cols from se by an expression involving its variables.
}
\examples{
# Arrange seq_se by gene_group and then gene_name
seq_se \%>\% arrange(row, gene_group, gene_name)
}
