% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{drop_metadata}
\alias{drop_metadata}
\title{Removes non-informative metadata}
\usage{
drop_metadata(se)
}
\arguments{
\item{se}{SummarizedExperiment to drop metadata for}
}
\value{
A SummarizedExperiment after redundant metadata has been dropped
}
\description{
For each of colData and rowData, removes the variables that have only 1 
unique value
}
\examples{
# as the se only contains time == 4 data, the time variable can be dropped
data(seq_se)
seq_se \%>\% filter(col, time == 4) \%>\% drop_metadata 
}
