% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{filter}
\alias{filter}
\alias{filter.SummarizedExperiment}
\title{Returns row of se with matching conditions}
\usage{
filter(se, ...)

\method{filter}{SummarizedExperiment}(se, ...)
}
\arguments{
\item{se}{SummarizedExperiment to subset}

\item{...}{Logical predicates defined in terms of the variables in .data. 
Multiple conditions are combined with & or ,. Only rows where the condition evaluates to TRUE are kept.}
}
\description{
Use filter() to choose rows/genes where conditions are true. Uses attached rowData
to find matches.
}
\examples{
#filter the rows of the example se with genes from the IL group
seq_se \%>\% filter(gene_group == "IL")
}
