% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{filter.SummarizedExperiment}
\alias{filter.SummarizedExperiment}
\title{Returns rows or cols of se with matching conditions}
\usage{
\method{filter}{SummarizedExperiment}(.data, axis, ...)
}
\arguments{
\item{.data}{SummarizedExperiment to subset}

\item{axis}{The axis to perform the operation on. Either row or col.}

\item{...}{Logical predicates defined in terms of the variables in .data. 
Multiple conditions are combined with & or ,. Only rows where the condition 
evaluates to TRUE are kept.}
}
\value{
A SummarizedExperiment after the filter operation
}
\description{
Use filter() to choose rows/genes where conditions are true. Uses attached 
metadata
to find matches.
}
\examples{
#filter the rows of the example se with genes from the IL group
data(seq_se)
seq_se \%>\% filter(row, gene_group == "IL")
}
