% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{get_col_data}
\alias{get_col_data}
\title{Gets column data from an se as a tibble}
\usage{
get_col_data(se)
}
\arguments{
\item{se}{SummarizedExperiment}
}
\description{
Gets column data from an se as a tibble
}
\examples{
get_col_data(seq_se)
}
