% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{get_row_data}
\alias{get_row_data}
\title{Gets row data from an se as a tibble}
\usage{
get_row_data(se)
}
\arguments{
\item{se}{SummarizedExperiment}
}
\description{
Gets row data from an se as a tibble
}
\examples{
get_row_data(seq_se)
}
