% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{options_from_coldata}
\alias{options_from_coldata}
\title{Gets available options for a particular column of the coldata of an se}
\usage{
options_from_coldata(se, colname)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{colname}{the column name (character) to extract options for}
}
\description{
Gets available options for a particular column of the coldata of an se
}
\examples{
options_from_coldata(seq_se, "site")
}
