% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{+.SummarizedExperiment}
\alias{+.SummarizedExperiment}
\title{Add two SummarizedExperiments}
\usage{
\method{+}{SummarizedExperiment}(se1, se2)
}
\arguments{
\item{se1, se2}{SummarizedExperiments for which se1 and se2 values will be 
added
Need to have the same rowData and colData}
}
\value{
A SummarizedExperiment after the add operation
}
\description{
Adds the values from each of the underlying assay values for 2 se's.
}
\examples{
#create subset se's for time == 0 and time == 4. Then add their values.
#remove time to make colData the same
data(seq_se)
t0 <- seq_se \%>\% filter(col, time == 0) \%>\% select(col, -time)
t4 <- seq_se \%>\% filter(col, time == 4) \%>\% select(col, -time)
t4 + t0
}
