% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{print_options}
\alias{print_options}
\title{Prints available options for the se}
\usage{
print_options(se)
}
\arguments{
\item{se}{SummarizedExperiment to show information for}
}
\description{
Creates a print of all possible assays, rowData and colData
}
\examples{
#show available options for the example se
print_options(seq_se)
}
