% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{slice}
\alias{slice_sample}
\alias{rename}
\alias{select}
\alias{arrange}
\alias{filter}
\alias{mutate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_sample}}}
}}

