% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{round.SummarizedExperiment}
\alias{round.SummarizedExperiment}
\title{Round the values of a SummarizedExperiments}
\usage{
\method{round}{SummarizedExperiment}(x, digits = 0)
}
\arguments{
\item{x}{SummarizedExperiment to round}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Rounds the values in all of the underlying assay matrices to the given number of digits
}
\examples{
# round the example se to 2 digits
seq_se \%>\% round(2)
}
