% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{sample_frac}
\alias{sample_frac}
\alias{sample_frac.SummarizedExperiment}
\title{Sample a fraction of total rows or cols from a se}
\usage{
sample_frac(se, axis, ...)

\method{sample_frac}{SummarizedExperiment}(se, axis, ...)
}
\arguments{
\item{se}{SummarizedExperiment to sample}

\item{axis}{The axis to perform the operation on. Either row or col.}

\item{...}{Additional arguments to sample_frac}
}
\description{
Selects random fraction of all rows or columns
}
\examples{
#Sample half of the genes from seq_se
seq_se \%>\% sample_frac(row, size = .5)
}
