% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{sample_frac_col}
\alias{sample_frac_col}
\title{Sample a fraction of cols from a se}
\usage{
sample_frac_col(tbl, size, replace = FALSE, weight = NULL, ...)
}
\arguments{
\item{tbl}{SummarizedExperiment to sample}

\item{size}{The fraction of cols to select}

\item{replace}{Sample with or without replacement?}

\item{weight}{Sampling weights. This must evaluate to a vector of non-negative 
numbers the same length as the input. Weights are automatically standardised to 
sum to 1.

This argument is automatically quoted and later evaluated in the context of the 
data frame. It supports unquoting. See vignette("programming") for an introduction 
to these concepts.}

\item{...}{ignored}
}
\description{
Selects a random fraction of cols
}
\examples{
#Sample half of the columns
seq_se \%>\% sample_frac_col(size = .5)
}
