% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{sample_n}
\alias{sample_n}
\alias{sample_n.SummarizedExperiment}
\title{Sample n rows or cols from a se}
\usage{
sample_n(se, axis, ...)

\method{sample_n}{SummarizedExperiment}(se, axis, ...)
}
\arguments{
\item{se}{SummarizedExperiment to sample}

\item{axis}{The axis to perform the operation on. Either row or col.}

\item{...}{Additional arguments to sample_n}
}
\description{
Selects random rows or columns
}
\examples{
#Sample 4 columns from seq_se with replacement
seq_se \%>\% sample_n(col, size = 4, replace = TRUE)
}
