% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{sample_n_row}
\alias{sample_n_row}
\title{Sample n rows from a se}
\usage{
sample_n_row(tbl, size, replace = FALSE, weight = NULL, ...)
}
\arguments{
\item{tbl}{SummarizedExperiment to sample}

\item{size}{The number of rows to select}

\item{replace}{Sample with or without replacement?}

\item{weight}{Sampling weights. This must evaluate to a vector of non-negative 
numbers the same length as the input. Weights are automatically standardised to 
sum to 1.

This argument is automatically quoted and later evaluated in the context of the 
data frame. It supports unquoting. See vignette("programming") for an introduction 
to these concepts.}

\item{...}{ignored}
}
\description{
Selects random rows
}
\examples{
#Sample 4 rows from seq_se with replacement
seq_se \%>\% sample_n_row(size = 4, replace = TRUE)
}
