% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{select}
\alias{select}
\alias{select.SummarizedExperiment}
\title{Returns columns of se with matching conditions}
\usage{
select(se, ...)

\method{select}{SummarizedExperiment}(se, ...)
}
\arguments{
\item{se}{SummarizedExperiment to subset}

\item{...}{Logical predicates defined in terms of the variables in .data. 
Multiple conditions are combined with & or ,. Only columns where the condition 
evaluates to TRUE are kept.}
}
\description{
Use select() to choose columns/samples where conditions are true. Uses attached colData
to find matches. Note that dplyr's filter syntax is used.
}
\examples{
#select the columns/samples of patient 1
seq_se \%>\% select(patient == 1)
#select the columns/samples where treatment is B and the site is either brain or skin
#' seq_se \%>\% select(treatment == "B", site \%in\% c("brain", "skin"))
}
