% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seq_se}
\alias{seq_se}
\title{Example sequencing SummarizedExperiment}
\format{
A SummarizedExperiment
}
\usage{
seq_se
}
\description{
Toy example SummarizedExperiment containing expression and copy number values for 
10 genes and 48 different conditions. The 10 genes are grouped in 3 gene groups.
The samples are taken from 3 different sites, which were treated with either 
treatment A or B for 0 or 4 hours for 4 different patients.
Use print_options(seq_se) to get an overview.
}
\keyword{datasets}
