% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{set_col_data}
\alias{set_col_data}
\title{Sets column data from an se}
\usage{
set_col_data(se, coldata)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{coldata}{data frame or tibble containing the coldata}
}
\description{
As this does not require inplace operations, this can typically be used
in a pipe
}
\examples{
new_cd <- get_col_data(seq_se) \%>\% tidyr::unite(time_treatment, c(time, treatment))
se_new_coldata <- set_col_data(seq_se, new_cd)
}
