% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{slice_sample.SummarizedExperiment}
\alias{slice_sample.SummarizedExperiment}
\title{Randomly selects rows or columns}
\usage{
\method{slice_sample}{SummarizedExperiment}(.data, axis, ..., n, prop)
}
\arguments{
\item{.data}{SummarizedExperiment to subset}

\item{axis}{The axis to perform the operation on. Either row or col.}

\item{...}{Arguments passed to dplyr::slice_sample}

\item{n}{See prop}

\item{prop}{Provide either n, the number of rows, or prop, the proportion of 
rows to select. If neither are supplied, n = 1 will be used. If n is greater 
than the number of rows in the group (or prop > 1), the result will be 
silently truncated to the group size. prop will be rounded towards zero to 
generate an integer number of rows. A negative value of n or prop will be 
subtracted from the group size. For example, n = -2 with a group of 5 rows 
will select 5 - 2 = 3 rows; prop = -0.25 with 8 rows will select 
8 * (1 - 0.25) = 6 rows.}
}
\value{
A SummarizedExperiment after the sample operation
}
\description{
Uses a slice operation to subset
}
\examples{
# subset 5 random columns
data(seq_se)
seq_se \%>\% slice_sample(col, n=5)
# randomly subset 20\% of rows
seq_se \%>\% slice_sample(row, prop=.2)
}
