% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{write_delim}
\alias{write_delim}
\title{Write a se to delim format}
\usage{
write_delim(se, path, delim = " ", assay_name = NULL)
}
\arguments{
\item{se}{SummarizedExperiment to round}

\item{path}{Path or connection to write to}

\item{delim}{Delimiter used to separate values}

\item{assay_name}{Name of the assay contained in se to write. Not required 
when only 1 assay is contained in the se}
}
\value{
The path to where the SummarizedExperiment was written
}
\description{
Rounds the values in all of the underlying assay matrices to the given 
number of digits
}
\examples{
# write the example se to delim file, defaulting to " " as delimiter
data(seq_se)
seq_se \%>\% round(3) \%>\% write_delim("out.txt", assay_name = "expression")
}
