% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{write_se_to_delim}
\alias{write_se_to_delim}
\title{Write a se to delim format}
\usage{
write_se_to_delim(se, assayname, path, delim = " ")
}
\arguments{
\item{se}{SummarizedExperiment to round}

\item{assayname}{Name of the assay contained in se to write}

\item{path}{Path or connection to write to}

\item{delim}{Delimiter used to separate values}
}
\description{
Rounds the values in all of the underlying assay matrices to the given number of digits
}
\examples{
# write the example se to delim file, defaulting to " " as delimiter
seq_se \%>\% round(3) \%>\% write_se_to_delim("expression", "out.txt")
}
