% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{write_se_to_tsv}
\alias{write_se_to_tsv}
\title{Write a se to tsv format}
\usage{
write_se_to_tsv(se, assayname, path)
}
\arguments{
\item{se}{SummarizedExperiment to round}

\item{assayname}{Name of the assay contained in se to write}

\item{path}{Path or connection to write to}
}
\description{
Rounds the values in all of the underlying assay matrices to the given number of digits
}
\examples{
# write the example se to tsv
seq_se \%>\% round(3) \%>\% write_se_to_tsv("expression", "out.tsv")
}
