% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{write_tsv}
\alias{write_tsv}
\title{Write a se to tsv format}
\usage{
write_tsv(se, path, assay_name = NULL)
}
\arguments{
\item{se}{SummarizedExperiment to round}

\item{path}{Path or connection to write to}

\item{assay_name}{Name of the assay contained in se to write. Not required when 
only 1 assay is contained in the se}
}
\description{
Rounds the values in all of the underlying assay matrices to the given number of digits
}
\examples{
# write the example se to tsv
seq_se \%>\% round(3) \%>\% write_tsv("out.tsv", "expression")
}
